requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bitsandbytes==0.38.1
black==23.1.0
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cloudpickle==2.2.1
colorlog==6.7.0
comm==0.1.3
commentjson==0.9.0
ConfigArgParse==1.5.3
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.33
dash==2.9.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
dearpygui==1.9.0
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.14.0
dill==0.3.6
easydict==1.10
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.3
fasttext==0.9.2
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
fonttools==4.38.0
freetype-py==2.3.0
frozenlist==1.3.3
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.2.2
gdown==4.7.1
ghalton==0.6.2
gitdb==4.0.10
GitPython==3.1.31
gnureadline==8.1.2
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.53.0
h5py==3.8.0
hickle==5.0.2
huggingface-hub==0.12.1
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.9
imgviz==1.7.2
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.23.6
jupyter-ydoc
jupyter_client
jupyter_core
jupyter_packaging
jupyter_server_fileid
jupyter_server_ydoc
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lark-parser==0.7.8
lazy_loader==0.2
lpips==0.1.4
lvis==0.5.3
lxml==4.9.2
mapbox-earcut==1.0.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
moviepy==1.0.3
mpmath==1.3.0
mss==7.0.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.18.0
open3d==0.16.0
openai==0.27.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.0
pluggy==1.2.0
portalocker==2.7.0
proglog==0.1.10
progressbar2==4.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pybullet==3.2.0
pybullet-planning==0.5.1
pycocotools==2.0.6
pycollada==0.7.2
pycparser==2.21
pyglet==1.5.27
Pygments==2.14.0
PyGObject==3.44.1
PyMCubes==0.1.4
PyOpenGL==3.1.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.19.3
pysdf==0.1.8
pyserial==3.5
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==3.5.2
python-version==0.0.2
pytorch3d==0.3.0
pytz==2023.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.2
recordclass==0.18.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
rsa==4.9
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.0
scikit-learn==1.0.2
scipy==1.7.3
-e git+ssh://git@github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
Send2Trash==1.8.0
sentencepiece==0.1.98
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
svg.path==6.2
sympy==1.10.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==1.15.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
terminado==0.17.1
thinplate @ git+https://github.com/cheind/py-thin-plate-spline@f6995795397118b7d0ac01aecd3f39ffbfad9dee
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.7
torch==1.13.1
torch-ema==0.3
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
trimesh==3.21.4
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.15
visualization==1.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
xxhash==3.2.0
y-py
yacs==0.1.8
yarl==1.9.2
ypy-websocket
zipp==3.15.0
# tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/@8d2536b8b324c998ff0ecec74e9d6a9c77bd45f3#subdirectory=bindings/torch # Try this if you are having tinycudann problems